added user role stuff...tried to fix buy and sell but no server

+++ application/models/Stock.php
@@ -14,6 +14,9 @@ class Stock extends MY_Model {
     }
 
     public function redoStocks($stockList){
+        if(is_null($stockList)){
+            return NULL;
+        }
          $this->load->database();
          $this->db->truncate('stocks');
 
@@ -31,7 +34,7 @@ class Stock extends MY_Model {
     }
     public function redoMovement($stockList){
          if(is_null($stockList)){
-            return;
+           return NULL;
         }
          $this->load->database();
          $this->db->truncate('movements');
@@ -52,7 +55,7 @@ class Stock extends MY_Model {
     public function redoTransactions($stockList){
         
         if(is_null($stockList)){
-            return;
+            return NULL;
         }
         
          $this->load->database();
